requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==1.4.45
Flask-JWT-Extended
Flask-CORS
python-dotenv
werkzeug

